requirements.txt
# SC2 API
burnysc2>=6.5.0
pysc2>=3.0.0
sc2reader>=1.8.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# RL and ML utilities
numpy>=1.24.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
